+++ packages/api/.gitignore
@@ -1,3 +1,3 @@
 *.d.ts
-src/**/*.js
+lib/**/*.js
 tests/*.js
