+++ packages/api/lib/rt/_vendor/aws-lambda/wrapApiClass.ts
