+++ packages/api/lib/rt/_vendor/nextjs/wrapApiClass.ts
